[fix] exclude video in image_cleaner

+++ src/util/image_cleaner.py
@@ -18,8 +18,9 @@ datas = Post.objects.all()
 
 if __name__ == '__main__':
     content_image_names = dict()
+    content_video_names = dict()
     title_image_names = dict()
-
+    
     for data in datas:
         title_image_names[str(data.image).split('/')[-1]] = 0
 
@@ -27,7 +28,11 @@ if __name__ == '__main__':
         images = soup.select('img')
         for image in images:
             content_image_names[image.get('src').split('/')[-1]] = 0
-
+        videos = soup.select('source')
+        for video in videos:
+            content_video_names[video.get('src').split('/')[-1]] = 0
+            print(content_video_names)
+    
     for(path, dir, files) in os.walk(TITLE_IMAGE_DIR):
         for filename in files:
             if not filename in title_image_names:
@@ -36,8 +41,8 @@ if __name__ == '__main__':
 
     for (path, dir, files) in os.walk(CONTENT_IMAGE_DIR):
         for filename in files:
-            if not filename in content_image_names:
+            if not filename in content_image_names and not filename in content_video_names:
                 print(path + '/' + filename)
                 os.remove(path + '/' + filename)
-
+    
     print('ALL DONE!')
